refactors configure tasks (#4)
This change rearranges the ios configure tasks and provides support for
loading the configuration from the terminal as opposed to the merge
function (which is still supported).  The configure tasks will also now
generate the diff post config push and use the diff output to determine
if a change is made.