update main.yml to include init tasks (#14)
This change updates the main.yml implementation to use the init.yaml set
of tasks in the includes folder.  The init.yaml is designed to be used
by an of the tasks in this role to check for basic values to be set in
order for the role to operate seamlessly.